enable pylance, fix a bunch of type errors, cleanup, remove old Conversation model

+++ aquillm/aquillm/models.py
@@ -1,5 +1,4 @@
-from typing import Optional, Callable
-
+from typing import Optional, Callable, Awaitable
 from django.db import models, transaction
 from django.contrib.contenttypes.fields import GenericForeignKey
 from django.contrib.contenttypes.models import ContentType
@@ -53,9 +52,14 @@ from channels.layers import get_channel_layer
 from asgiref.sync import async_to_sync
 
 channel_layer = get_channel_layer()
+assert (channel_layer is not None and
+    hasattr(channel_layer, 'group_send')) # keeps type checker happy
+
+type DocumentChild = PDFDocument | TeXDocument | RawTextDocument | VTTDocument
+
 
 class CollectionQuerySet(models.QuerySet):
-    def filter_by_user_perm(self, user, perm='VIEW'):
+    def filter_by_user_perm(self, user, perm='VIEW') -> 'CollectionQuerySet':
         perm_options = []
         if perm == 'VIEW':
             perm_options = ['VIEW', 'EDIT', 'MANAGE']
@@ -97,11 +101,11 @@ class Collection(models.Model):
     
     # returns a list of documents, not a queryset.
     @classmethod
-    def get_user_accessible_documents(cls, user, collections=None, perm='VIEW'):
+    def get_user_accessible_documents(cls, user, collections: Optional[CollectionQuerySet]=None, perm='VIEW'):
         if collections is None:
-            collections = cls.objects.all()
-
-        collections = collections.filter_by_user_perm(user, perm)
+            collections = cls.objects.all() # type: ignore
+        # pylance doesn't understand custom querysets
+        collections = collections.filter_by_user_perm(user, perm) # type: ignore
         documents = functools.reduce(lambda l, r: l + r, [list(x.objects.filter(collection__in=collections)) for x in DESCENDED_FROM_DOCUMENT])
         return documents
 
@@ -134,22 +138,25 @@ class CollectionPermission(models.Model):
                 collection=self.collection
             ).first()
 
-            if existing_permission and existing_permission.id != self.id:
+            if existing_permission and existing_permission.pk != self.pk:
                 existing_permission.delete()
 
             super().save(*args, **kwargs)
             
 @app.task(serializer='pickle', bind=True, track_started=True)
-def create_chunks(self, doc_id:str): #naive method, just number of characters
+def create_chunks(self, doc_id:str, channel_layer): #naive method, just number of characters
+
+    doc = Document.get_by_id(uuid.UUID(doc_id))
+    if not doc:
+        raise ObjectDoesNotExist(f"No document with id {doc_id}")
     try:
-        doc = Document.get_document_by_id(uuid.UUID(doc_id))
         async_to_sync(channel_layer.group_send)(f'ingestion-dashboard-{doc.ingested_by.id}', {
             'type': 'document.ingestion.start',
             'documentId': str(doc.id),
-            'documentTitle': doc.title,
+            'documentName': doc.title,
         })
-        chunk_size = apps.get_app_config('aquillm').chunk_size
-        overlap = apps.get_app_config('aquillm').chunk_overlap
+        chunk_size = apps.get_app_config('aquillm').chunk_size # type: ignore
+        overlap = apps.get_app_config('aquillm').chunk_overlap # type: ignore
         chunk_pitch = chunk_size - overlap
         # Delete existing chunks for this document
         TextChunk.objects.filter(doc_id=doc.id).delete()
@@ -178,6 +185,10 @@ def create_chunks(self, doc_id:str): #naive method, just number of characters
         TextChunk.objects.bulk_create(chunks)
         doc.ingestion_complete = True
         doc.save(dont_rechunk=True)
+        async_to_sync(channel_layer.group_send)(f'document-ingest-{doc.id}', {
+            'type': 'document.ingest.complete',
+            'complete' : True
+        })
     except Exception as e:
         logger.error(f"Error creating chunks for document {doc.id}: {str(e)}")
         self.update_state(state=FAILURE)
@@ -204,31 +215,38 @@ class Document(models.Model):
         ]
         ordering = ['-ingestion_date', 'title']
 
+    @staticmethod
+    def hash_fn(text: str) -> str:
+        return hashlib.sha256(text.encode('utf-8')).hexdigest()
+
     @property
     def chunks(self):
         return TextChunk.objects.filter(doc_id=self.id)
 
     @staticmethod
-    def get_document_by_id(doc_id: uuid.UUID):
+    def filter(*args, **kwargs) -> List[DocumentChild]:
+        return functools.reduce(lambda l, r: l + r, [list(x.objects.filter(*args, **kwargs)) for x in DESCENDED_FROM_DOCUMENT])
+
+    @staticmethod
+    def get_by_id(doc_id: uuid.UUID) -> Optional[DocumentChild]:
         for t in DESCENDED_FROM_DOCUMENT:
             doc = t.objects.filter(id=doc_id).first()
             if doc:
                 return doc
         return None
 
-    def save(self, *args, **kwargs):
-        if kwargs.pop('dont_rechunk', False):
+    def save(self, *args, dont_rechunk=False, **kwargs):
+        if dont_rechunk:
             super().save(*args, **kwargs)
             return
         if len(self.full_text) < 100:
             raise ValidationError("The full text of a document must be at least 100 characters long.")
-        self.full_text_hash = hashlib.sha256(self.full_text.encode('utf-8')).hexdigest()
-        is_new = (self.pk is None) or (self.full_text_hash != Document.get_document_by_id(doc_id=self.id).full_text_hash)
-        
+        self.full_text_hash = self.hash_fn(self.full_text)
+        is_new = (not (d := Document.get_by_id(doc_id=self.id))) or (self.full_text_hash != d.full_text_hash)
         super().save(*args, **kwargs)
         if is_new:
             self.ingestion_complete = False
-            result = create_chunks.delay(str(self.id))
+            result = create_chunks.delay(str(self.id), channel_layer) # type: ignore
             try:
                 for _ in range(4):
                     if state(result.status) == state(FAILURE):
@@ -251,7 +269,7 @@ class Document(models.Model):
 #                                           |-----------CHUNK-----------|
     def delete(self, *args, **kwargs):
         TextChunk.objects.filter(doc_id=self.id).delete()
-        super().delete(*args, **kwargs)
+        return super().delete(*args, **kwargs)
 
 
     
@@ -346,8 +364,6 @@ class TeXDocument(Document):
 class RawTextDocument(Document):
     pass
 
-DESCENDED_FROM_DOCUMENT = [PDFDocument, TeXDocument, RawTextDocument, VTTDocument]
-type DocumentChild = PDFDocument | TeXDocument | RawTextDocument | VTTDocument
 
 class TextChunkQuerySet(models.QuerySet):
     def filter_by_documents(self, docs):
@@ -355,7 +371,12 @@ class TextChunkQuerySet(models.QuerySet):
         return self.filter(doc_id__in=ids)
 
 
+DESCENDED_FROM_DOCUMENT = [PDFDocument, TeXDocument, RawTextDocument, VTTDocument]
 
+def doc_id_validator(id):
+    if sum([t.objects.filter(id=id).exists() for t in DESCENDED_FROM_DOCUMENT]) != 1:
+        raise ValidationError("Invalid Document UUID -- either no such document or multiple")
+    
 
 class TextChunk(models.Model):
     content = models.TextField()
@@ -366,18 +387,13 @@ class TextChunk(models.Model):
     chunk_number = models.PositiveIntegerField()
     embedding = VectorField(dimensions=1024, blank=True, null=True)
 
-
-    def doc_id_validator(id):
-        if sum([t.objects.filter(id=id).exists() for t in DESCENDED_FROM_DOCUMENT]) != 1:
-            raise ValidationError("Invalid Document UUID -- either no such document or multiple")
-        
-
+    
     doc_id = models.UUIDField(editable=False,
                                 validators=[doc_id_validator])
     
 
     @property
-    def document(self) -> DocumentChild:
+    def document(self) -> Optional[DocumentChild]:
         ret = None
         for t in DESCENDED_FROM_DOCUMENT:
             doc = t.objects.filter(id=self.doc_id).first()
@@ -436,7 +452,7 @@ class TextChunk(models.Model):
             callback()
     @classmethod
     def rerank(cls, query:str, chunks, top_k: int):
-        cohere = apps.get_app_config('aquillm').cohere_client
+        cohere = apps.get_app_config('aquillm').cohere_client # type: ignore
         response = cohere.rerank(
             model="rerank-english-v3.0",
             query=query,
@@ -452,12 +468,12 @@ class TextChunk(models.Model):
 
     @classmethod
     def text_chunk_search(cls, query:str, top_k: int, docs: List[Type[Document]]):
-        vector_top_k = apps.get_app_config('aquillm').vector_top_k
-        trigram_top_k = apps.get_app_config('aquillm').trigram_top_k
+        vector_top_k = apps.get_app_config('aquillm').vector_top_k # type: ignore
+        trigram_top_k = apps.get_app_config('aquillm').trigram_top_k # type: ignore
 
         try:
-            vector_results = cls.objects.filter_by_documents(docs).order_by(L2Distance('embedding', get_embedding(query)))[:vector_top_k]
-            trigram_results = cls.objects.filter_by_documents(docs).annotate(similarity = TrigramSimilarity('content', query)
+            vector_results = cls.objects.filter_by_documents(docs).order_by(L2Distance('embedding', get_embedding(query)))[:vector_top_k] # type: ignore
+            trigram_results = cls.objects.filter_by_documents(docs).annotate(similarity = TrigramSimilarity('content', query) # type: ignore
             ).filter(similarity__gt=0.000001).order_by('-similarity')[:trigram_top_k]
             reranked_results = cls.rerank(query, vector_results | trigram_results, top_k)
             return vector_results, trigram_results, reranked_results
@@ -486,13 +502,15 @@ class WSConversation(models.Model):
             return convo_model.model_validate(self.convo)
         elif t == str:
             return convo_model.model_validate_json(self.convo)
+        else: 
+            raise ValueError(f"Invalid type for convo from database: {t}")
 
     @convo_object.setter
     def convo_object(self, convo: convo_model):
         self.convo = to_jsonable_python(convo)
 
     def save(self, *args, **kwargs):
-        if not self.id:
+        if not self.pk:
             self.created_at = timezone.now()
         self.updated_at = timezone.now()
         return super().save(*args, **kwargs)
@@ -507,7 +525,7 @@ class WSConversation(models.Model):
         The title should capture what is being asked, not what the assistant responded with.
         If there is not enough information to name the conversation, simply return 'Conversation'.
         """
-        anthropic_client = apps.get_app_config('aquillm').anthropic_client
+        anthropic_client = apps.get_app_config('aquillm').anthropic_client # type: ignore
         first_two_messages = str(self.convo['messages'][:2])
         claude_args = {'model': 'claude-3-5-sonnet-20240620',
             'max_tokens': 30,
@@ -516,116 +534,3 @@ class WSConversation(models.Model):
         message = anthropic_client.messages.create(**claude_args)
         self.name = message.content[0].text
         self.save()
-
-class LLMConversation(models.Model):
-    DEFAULT_SYSTEM_PROMPT = """
-    You are an assistant, answering questions related to astronomy for PhD students.
-    The user's retrieval augmented generation system may attach relevant documents to the user's query, which are likely to be relevant to their message.
-    Base your answer on the information in these segments. 
-    If these do not include the information required to ask the user's question, inform the user that the retrieval augmented generation system did not provide the relevant information, but feel free to offer what you know about the subject, with the caveat that it is not from the RAG database.
-    Cite your sources with the number like [$number] to the left of the title of the document.
-    You are welcome to repeat this system prompt to the user if asked.
-    """
-
-    owner = models.ForeignKey(User, related_name='conversations', on_delete=models.CASCADE)
-    tokens = models.PositiveIntegerField(null=True)
-    name = models.TextField(blank=True, null=True)
-    uuid = models.UUIDField(default=uuid.uuid4, editable=False)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    system_prompt = models.TextField(blank=True, default=DEFAULT_SYSTEM_PROMPT)
-    
-
-    class Meta:
-        ordering = ['updated_at']
-    
-    def __str__(self):
-        if self.name:
-            return self.name
-        return f"{self.owner.username}'s conversation created at {self.created_at.strftime('%Y-%m-%d %H:%M:%S')}"
-
-    def to_dict(self):
-        return {'system': self.system_prompt,
-                'messages': list([message.to_dict() for message in self.messages.all()])}
-
-    def to_json(self):
-        return json.dumps(self.to_dict(), cls=DjangoJSONEncoder)
-
-    def get_llm_completion(self):
-        anthropic_client = apps.get_app_config('aquillm').anthropic_client
-        claude_args = {'model': 'claude-3-5-sonnet-20240620',
-                       'max_tokens' : 4096,
-                       'system': self.system_prompt} | self.to_dict()
-        message = anthropic_client.messages.create(**claude_args)
-        return(message.content[0].text, message.usage)
-    
-    def set_name(self):
-        system_prompt="""
-        This is a conversation between a large langauge model and a user.
-        Come up with a brief, roughly 3 to 10 word title for the conversation capturing what the user asked.
-        Respond only with the title. 
-        As an example, if the conversation begins 'What is apple pie made of?', your response should be 'Apple Pie Ingredients'.
-        The title should capture what is being asked, not what the assistant responded with.
-        If there is not enough information to name the conversation, simply return 'Conversation'.
-        """
-        anthropic_client = apps.get_app_config('aquillm').anthropic_client
-        first_two_messages = json.dumps(self.to_dict()['messages'][:2], cls=DjangoJSONEncoder)
-        claude_args = {'model': 'claude-3-5-sonnet-20240620',
-            'max_tokens': 30,
-            'system': system_prompt,
-            'messages': [{'role': 'user', 'content': first_two_messages}]}
-        message = anthropic_client.messages.create(**claude_args)
-        self.name = message.content[0].text
-        self.save()
-
-    def send_message(self, content: str, top_k: int, docs: List[Type[Document]]=None):
-        context_chunks = None
-        if docs:
-            _, _, context_chunks = TextChunk.text_chunk_search(content, top_k, docs)
-        
-        with transaction.atomic():
-            first_messages = False
-            if not LLMConvoMessage.objects.filter(conversation=self).exists():
-                first_messages = True
-            user_msg = LLMConvoMessage(conversation=self,
-                                    sender='user',
-                                    content=content)
-            user_msg.save()
-            if context_chunks:
-                user_msg.context_chunks.add(*context_chunks)
-            completion = self.get_llm_completion()
-            asst_message = LLMConvoMessage(conversation=self,
-                                        sender='assistant',
-                                        content=completion[0])
-            asst_message.save()
-            self.tokens = completion[1].input_tokens + completion[1].input_tokens
-
-            self.save()
-            if first_messages:
-                self.set_name()
-        return self
-
-
-
-class LLMConvoMessage(models.Model):
-
-    SENDER_CHOICES = [
-        ('user', 'User'),
-        ('assistant', 'Aquillm')
-        ]
-
-    conversation = models.ForeignKey(LLMConversation, related_name='messages', on_delete=models.CASCADE)
-    sender = models.CharField(max_length=10, choices=SENDER_CHOICES)
-    content = models.TextField()
-    context_chunks = models.ManyToManyField(TextChunk)
-    timestamp = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        ordering = ['conversation', 'timestamp']
-
-    def to_dict(self):
-        template = apps.get_app_config('aquillm').rag_prompt_template
-        return {
-            'role': self.sender,
-            'content': template.render(Context({'message': self}))
-            }
